+++ packages/common-all/src/dnode.ts
@@ -57,6 +57,7 @@ export class DNodeUtils {
       children,
       body,
       data,
+      contentHash,
       vault,
     } = _.defaults(opts, {
       updated: Time.now().toMillis(),
@@ -87,6 +88,7 @@ export class DNodeUtils {
       children,
       body,
       data,
+      contentHash,
     };
 
     // don't include optional props
@@ -750,9 +752,10 @@ export class NoteUtils {
   }
 
   /**
-   * Add props from @param noteHydrated to @param noteRaw
-   * @param param0
-   * @returns
+   * Add {@link DNodeProps["parent"]} and {@link DNodeProps["children"]} from {@link noteHydrated} to {@link noteRaw}
+   * @param noteRaw - note for other fields
+   * @param noteHydrated - note to get {@link DNodeProps["parent"]} and {@link DNodeProps["children"]}  properties from
+   * @returns Merged Note object
    */
   static hydrate({
     noteRaw,