+++ packages/plugin-core/src/features/FrontmatterFoldingRangeProvider.ts
@@ -1,5 +1,8 @@
-import { NoteUtils } from "@dendronhq/common-all";
-import vscode from "vscode";
+import { MDUtilsV5, ProcMode } from "@dendronhq/engine-server";
+import vscode, { FoldingRangeKind } from "vscode";
+import visit from "unist-util-visit";
+import _ from "lodash";
+import { VSCodeUtils } from "../utils";
 
 export default class FrontmatterFoldingRangeProvider
   implements vscode.FoldingRangeProvider
@@ -12,14 +15,25 @@ export default class FrontmatterFoldingRangeProvider
   public async provideFoldingRanges(
     document: vscode.TextDocument
   ): Promise<vscode.FoldingRange[]> {
-    const content = document.getText();
-    const fmMatch = content.match(NoteUtils.RE_FM);
-    if (!fmMatch) {
-      return [];
-    }
-    const fmContent = fmMatch[0];
-    // we know fmContent cannot be null since fmMatch exists.
-    const fmContentLength = fmContent.match(/^/gm)!.length;
-    return [new vscode.FoldingRange(0, fmContentLength)];
+    const proc = MDUtilsV5.procRemarkParse({
+      mode: ProcMode.NO_DATA,
+      parseOnly: true,
+    });
+    const parsed = proc.parse(document.getText());
+    let range: vscode.FoldingRange | undefined;
+    visit(parsed, ["yaml"], (node) => {
+      if (_.isUndefined(node.position)) return false; // Should never happen
+      range = new vscode.FoldingRange(
+        VSCodeUtils.point2VSCodePosition(node.position.start).line,
+        VSCodeUtils.point2VSCodePosition(node.position.end).line,
+        FoldingRangeKind.Region
+      );
+
+      // Found the frontmatter already, stop traversing
+      return false;
+    });
+
+    if (_.isUndefined(range)) return [];
+    return [range];
   }
 }