+++ packages/plugin-core/src/fileWatcher.ts
@@ -7,15 +7,20 @@ import {
   WorkspaceOpts,
 } from "@dendronhq/common-all";
 import { file2Note, string2Note } from "@dendronhq/common-server";
-import { HistoryService, LinkUtils } from "@dendronhq/engine-server";
+import {
+  HistoryService,
+  LinkUtils,
+  WorkspaceUtils,
+} from "@dendronhq/engine-server";
 import _ from "lodash";
 import { AnchorUtils } from "@dendronhq/engine-server";
 import path from "path";
 import * as vscode from "vscode";
 import { Logger } from "./logger";
 import { DendronWorkspace, getWS } from "./workspace";
+import { ShowPreviewV2Command } from "./commands/ShowPreviewV2";
 
-export class VaultWatcher {
+export class FileWatcher {
   public watchers: { vault: DVault; watcher: vscode.FileSystemWatcher }[];
   /**
    * Should watching be paused
@@ -58,17 +63,16 @@ export class VaultWatcher {
   }
 
   /**
-   * Update note links
-   * @param uri
-   * @returns
+   * Update note on detected changes
+   *
+   * @returns {@link NoteProps} if changed, `undefined` otherwise
    */
   async onDidChange(uri: vscode.Uri) {
-    const ctx = "VaultWatcher:onDidChange";
-    this.L.debug({ ctx, uri: uri.fsPath });
+    const ctx = "FileWatcher:onDidChange";
     if (this.pause) {
       return;
     }
-    this.L.info({ ctx, uri });
+    this.L.info({ ctx, uri: uri.fsPath });
     const eclient = DendronWorkspace.instance().getEngine();
     const fname = path.basename(uri.fsPath, ".md");
     const doc = await vscode.workspace.openTextDocument(uri);
@@ -82,13 +86,21 @@ export class VaultWatcher {
       wsRoot: DendronWorkspace.wsRoot(),
       fsPath: uri.fsPath,
     });
-    let note = string2Note({ content, fname, vault });
     const noteHydrated = NoteUtils.getNoteByFnameV5({
       fname,
       vault,
       notes: eclient.notes,
       wsRoot: DendronWorkspace.wsRoot(),
     }) as NoteProps;
+    if (!WorkspaceUtils.noteContentChanged({ content, note: noteHydrated })) {
+      this.L.debug({
+        ctx,
+        uri: uri.fsPath,
+        msg: "note content unchanged, ignoring",
+      });
+      return;
+    }
+    let note = string2Note({ content, fname, vault, calculateHash: true });
     note = NoteUtils.hydrate({ noteRaw: note, noteHydrated });
     const links = LinkUtils.findLinks({ note, engine: eclient });
     note.links = links;
@@ -98,11 +110,13 @@ export class VaultWatcher {
     );
     note.anchors = anchors;
     this.L.info({ ctx, fname, msg: "exit" });
-    return await eclient.updateNote(note);
+    const noteClean = await eclient.updateNote(note);
+    ShowPreviewV2Command.refresh(noteClean);
+    return noteClean;
   }
 
   async onDidCreate(uri: vscode.Uri): Promise<void> {
-    const ctx = "VaultWatcher:onDidCreate";
+    const ctx = "FileWatcher:onDidCreate";
     if (this.pause) {
       this.L.info({ ctx, uri, msg: "paused" });
       return;
@@ -162,13 +176,13 @@ export class VaultWatcher {
         throw err;
       }
     } finally {
-      VaultWatcher.refreshTree();
+      FileWatcher.refreshTree();
       this.L.debug({ ctx, uri, msg: "refreshTree" });
     }
   }
 
   async onDidDelete(uri: vscode.Uri) {
-    const ctx = "VaultWatcher:onDidDelete";
+    const ctx = "FileWatcher:onDidDelete";
     if (this.pause) {
       return;
     }
@@ -213,7 +227,7 @@ export class VaultWatcher {
         // this.L.error({ ctx, err: JSON.stringify(err) });
       }
     } finally {
-      VaultWatcher.refreshTree();
+      FileWatcher.refreshTree();
     }
   }